update census to use layout
